[profile.d] Update pager configuration
Rename `less.sh` to `pager.sh` because the fact that I use `less(1)` as
a pager is merely an implementation detail.

Also, the `hash` invokation is broken on FreeBSD, as it returns `0` even
if its argument does not exist. This is because
[POSIX underspecified `hash(1)`](https://stackoverflow.com/questions/592620/how-can-i-check-if-a-program-exists-from-a-bash-script).
Switch to `command(1)` instead, as described in the previous link.