[profile.d] Replace lowercase environment variables
Replace lowercase environment variables with uppercase versions.

On Plan 9, environment variables are usually lowercase. While I like
many of the Plan 9 conventions, environment variables are expected to be
uppercase on Unix-like systems. In the worst case, an environment
variable might be picked up by a script which did not expect environment
variables to be all lowercase.

Signed-off-by: Florian Limberger <flo@purplekraken.com>