[Xresources] Configure XTerm scrollbar
Make the XTerm scrollbar similar to the Plan 9 scrollbar (as used by
e.g. sam):

* Scrollbar is on the left (already the default)
* Use a solid colour for the handle (thumb in XTerm parlance)
* Reverse the default scroll direction on mouse clicks:
  left button scrolls down, right button scrolls up

XTerm only uses two primary colours, the foreground and the background,
so the scrollbar uses the foreground colour. Also it is hidden by
default, since I currently mostly use tmux(1) inside, which has no use
for it. But a keybinding to quickly toggle it is added to make it more
convenient.