[git] Add sup alias and remove local config
The inclusion of gitconfig.local was an attempt to simplify switching
between my work and my personal identity. The plan was to have the per-
sonal identity the default, because the dotfiles are used on my perso-
nal machines too, and have the professional identity be a local over-
ride. Unfortunately, the repo utility used by the AOSP only checks the
gitconfig file and does not do any include handling. Therefore, this
file is not in use anymore and the email is overwritten locally.

Add the sup alias to simplify git submodule handling.