[nvim] Enable global status line
Use only a single status line for all open windows. The main use is to
have more space in the status line with multiple vertical splits.

Note that setting the globalstatus option in the lualine configuration
did not work, but setting vim.opt.laststatus to 3 did. This requires
neovim versions > 0.7, i.e. the neovim PPA on Ubuntu 22.04.